+++ packages/spacecat-shared-rum-api-client/src/functions/form-vitals.js
@@ -12,12 +12,12 @@
 
 import { DataChunks, facetFns, facets } from '@adobe/rum-distiller';
 import trafficAcquisition from './traffic-acquisition.js';
-import { generateKey, DELIMITER, loadBundles } from '../utils.js';
+import { DELIMITER, generateKey, loadBundles } from '../utils.js';
 
 const { checkpointSource } = facetFns;
 const { checkpoint: checkpointFacet, url: urlFacet } = facets;
 const METRICS = ['formview', 'formengagement', 'formsubmit'];
-const CHECKPOINTS = ['viewblock', 'click', 'fill', 'formsubmit', 'navigate', 'viewmedia'];
+const CHECKPOINTS = ['viewblock', 'click', 'fill', 'formsubmit', 'navigate', 'viewmedia', 'experiment'];
 const KEYWORDS_TO_FILTER = ['search'];
 
 function initializeResult(url) {
@@ -39,7 +39,7 @@ function filterEvents(bundles) {
         return false;
       }
 
-      if (event.checkpoint === 'navigate') {
+      if (event.checkpoint === 'navigate' || event.checkpoint === 'experiment') {
         return true;
       }
 
@@ -81,6 +81,18 @@ const metricFns = {
 
 function findByUrl(formVitals, url) {
   return Object.values(formVitals).find((item) => item.url === url);
+}
+
+function getAllURLWithExperiment(bundles) {
+  const dataChunks = new DataChunks();
+  loadBundles(bundles, dataChunks);
+  const { checkpoint } = facets;
+  dataChunks.addFacet('checkpoint', checkpoint);
+  return new Set(
+    dataChunks.facets.checkpoint
+      .filter((cp) => cp.value === 'experiment')
+      .flatMap((_checkpoint) => _checkpoint.entries.map((_entry) => _entry.url)),
+  );
 }
 
 function populateFormsInternalNavigation(bundles, formVitals) {
@@ -172,6 +184,10 @@ function findFormCTAWithinPage(bundles, formVitals) {
 
 function containsFormVitals(row) {
   return METRICS.some((metric) => Object.keys(row[metric]).length > 0);
+}
+
+function isUnderExperiment(row, experimentUrls) {
+  return experimentUrls.has(row.url);
 }
 
 function getParentPageVitalsGroupedByIFrame(bundles, dataChunks, iframeParentMap) {
@@ -309,10 +325,14 @@ function handler(bundles) {
     iframeParentMapWithoutDuplicates,
   );
 
+  const experimentUrls = getAllURLWithExperiment(bundlesWithFilteredEvents);
+
   // populate internal navigation data
   populateFormsInternalNavigation(bundles, formVitals);
   // filter out pages with no form vitals
-  const filteredFormVitals = Object.values(formVitals).filter(containsFormVitals);
+  const filteredFormVitals = Object.values(formVitals).filter(
+    (formVital) => containsFormVitals(formVital) && !isUnderExperiment(formVital, experimentUrls),
+  );
   findFormCTAForInternalNavigation(bundles, filteredFormVitals);
   const formVitalsWithCTA = findFormCTAWithinPage(bundles, filteredFormVitals);