+++ packages/spacecat-shared-content-client/CHANGELOG.md
@@ -1,3 +1,10 @@
+# [@adobe/spacecat-shared-content-client-v1.7.4](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-content-client-v1.7.3...@adobe/spacecat-shared-content-client-v1.7.4) (2025-05-31)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#779](https://github.com/adobe/spacecat-shared/issues/779)) ([07f8cce](https://github.com/adobe/spacecat-shared/commit/07f8cce73e33bfb9c61fe14f2ef28012b872437d))
+
 # [@adobe/spacecat-shared-content-client-v1.7.3](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-content-client-v1.7.2...@adobe/spacecat-shared-content-client-v1.7.3) (2025-05-26)